refactor: Migrate Storage engine to directory engine.
